# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a RESTful API for an E-commerce platform built with Go.",
        "title": "GO Commerce API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/products": {
            "get": {
                "description": "Get all the products from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get All Products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Create Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Product Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid Input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "description": "Get a single product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get One Products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "description": "Get all the users from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get All Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/email/{email}": {
            "get": {
                "description": "Get a single user from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Single Users By Email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User Email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/id/{id}": {
            "get": {
                "description": "Get a single user from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Single Users by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateProductRequest": {
            "description": "A product contains details like name, description, price, stock, category, and image URL.",
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 3
                },
                "description": {
                    "type": "string",
                    "example": "A brand-new smartphone with advanced features"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/smartphone.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "Nokia"
                },
                "price": {
                    "type": "number",
                    "example": 799.99
                },
                "stock": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "category_id": {
                    "description": "ID of the category the product belongs to",
                    "type": "integer",
                    "example": 2
                },
                "created_at": {
                    "description": "Timestamp when the product was created",
                    "type": "string",
                    "example": "2023-08-01T12:00:00Z"
                },
                "description": {
                    "description": "Product Description",
                    "type": "string",
                    "example": "A high-performance laptop"
                },
                "id": {
                    "description": "Product ID",
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "description": "URL of the product image",
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "name": {
                    "description": "Product Name",
                    "type": "string",
                    "example": "Lenovo 96"
                },
                "price": {
                    "description": "Price of the Product",
                    "type": "number",
                    "example": 999.99
                },
                "stock": {
                    "description": "Number of Stocks",
                    "type": "integer",
                    "example": 50
                }
            }
        },
...
        "models.UpdateProductRequest": {
            "description": "A product contains details like name, description, price, stock, category, and image URL.",
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 3
                },
                "description": {
                    "type": "string",
                    "example": "A brand-new smartphone with advanced features"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/smartphone.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "Nokia"
                },
                "price": {
                    "type": "number",
                    "example": 799.99
                },
                "stock": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Optional / new fields",
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "user or admin",
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, suspended",
                    "type": "string"
                }
            }
        }
    }
}